Allow parallel workers to get combo cid and distributed snapshot from DSM
When using parallel workers to create an HNSW index for heap-partitioned tables, pgvector reports an error: ERROR: Reader qExec tried to allocate a Combo Command Id (combocid.c:241) .

The error occurs because the function ParallelWorkerMain calls RestoreComboCIDState to restore the Combo CID from the dynamic shared memory segment. The Combo CID is stored in the dynamic shared memory segment by the function InitializeParallelDSM. RestoreComboCIDState invokes GetComboCommandId to build a local hash table, which was originally assumed to work only for QE writers. This mismatch leads to the error.

In practice, however, parallel workers (with Gp_role = GP_ROLE_EXECUTE and Gp_is_writer = false) should also be able to obtain a Combo CID, this is consistent with how PostgreSQL parallel workers behave.

The error occurs for partitioned heap tables but not for non-partitioned heap tables: for parallel index creation on partitioned heap tables, Combo CIDs are generated during the handling of parent and child tables; for non-partitioned heap tables, no Combo CIDs are generated, so GetComboCommandId is not called (since num_elements is zero).

We can determine if a process is a parallel worker via MyProc->lockGroupLeader: parallel workers have a lock group leader, which is the QE process that requests the postmaster to fork them. After a parallel worker is forked, it calls BecomeLockGroupMember to join the group locking mechanism.

In this PR, we allow parallel workers to retrieve Combo CIDs from the dynamic shared memory segment and build a local hash table for them. However, they will not write Combo CIDs back to the DSM (Dynamic Shared Memory).

Additionally, this PR serializes and deserializes distributed snapshots for parallel workers, enabling them to obtain the consistent snapshot from their leader QE process.

This PR will not affect the built-in indexes of WarehousePG. It will take effect for third-party indexes that use parallel workers for index building.