fix ORCA crash of percentile agg (#92)
This PR is trying to fix #91

---

When ORCA generates multistage aggregate plans for `percentile_disc()`
and `percentile_cont()`, it invokes the WHPG-specific transition
functions `gp_percentile_disc_transition` and `gp_percentile_cont_transition`.

These transition functions are non-strict, meaning that for the very first input 
row the incoming transition state (`transValue`) is not initialized by the executor
—it is passed in as Datum 0, representing a “null” state.

For pass-by-value data type, this does not cause a problem because the 
executor directly copies the returned value into the transition slot.

For pass-by-reference data type (e.g., numeric, interval), the executor treats 
the returned value as a pointer and will:

- keep the pointer to the transition value,
- free the old pointer if a new one is returned.

However, the transition functions incorrectly assumed that the transition state 
for the first row was already initialized (as would be true for strict functions). 
As a result, they simply returned the previous state—which was still Datum 0, 
instead of treating it as a null and setting it to the first row’s input value.

This led the executor to treat a null pointer as a valid address for pass-by-ref 
data type, causing a crash during aggregate execution.